Improve hot reloading
Since redux v2.0 hot reloading isn't automatically anymore, so there
was necessary to make a couple of configurations to avoid warning
messages.

The store and history were moved to another file because both need to
be created in a file which is part of the entry point otherwise the
warning message persist.